+++ app/client/admin/js/ovPub/ConfigurationController.js
@@ -12,9 +12,11 @@
   utilService,
   $filter,
   groups,
-  users
+  users,
+  properties
   ) {
     var publishMedias = publishConf.data.publishMedias;
+    var publishTls = publishConf.data.publishTls;
 
     $scope.rights = {};
     $scope.rights.edit = $scope.checkAccess('publish-manage-publish-config');
@@ -56,6 +58,30 @@
       isFormSaving: false
     };
 
+    // TLS settings
+    $scope.tlsSettings = {
+      model: {
+        properties: (publishTls && publishTls.properties) || []
+      },
+      fields: [
+        {
+          key: 'properties',
+          type: 'editableTags',
+          wrapper: ['editableWrapper', 'bootstrapLabel', 'bootstrapHasError'],
+          templateOptions: {
+            label: $filter('translate')('PUBLISH.CONFIGURATION.TLS_PROPERTIES'),
+            availableOptions: utilService.buildSelectOptions(properties.data.entities)
+          }
+        }
+      ],
+      options: {
+        formState: {
+          showForm: $scope.rights.edit
+        }
+      },
+      isFormSaving: false
+    };
+
     /**
      * Saves medias settings.
      */
@@ -72,10 +98,27 @@
         $scope.mediasSettings.isFormSaving = false;
       });
     };
+
+    /**
+     * Saves TLS settings.
+     */
+    $scope.saveTlsSettings = function() {
+      $scope.tlsSettings.isFormSaving = true;
+
+      return publishService.saveTlsSettings({
+        properties: $scope.tlsSettings.model.properties
+      }).then(function() {
+        $scope.tlsSettings.isFormSaving = false;
+        $scope.$emit('setAlert', 'success', $filter('translate')('CORE.UI.SAVE_SUCCESS'), 4000);
+      }).catch(function() {
+        $scope.tlsSettings.isFormSaving = false;
+      });
+    };
+
   }
 
   app.controller('ConfigurationController', ConfigurationController);
   ConfigurationController.$inject = [
-    '$scope', 'publishConf', 'publishService', 'utilService', '$filter', 'groups', 'users'];
+    '$scope', 'publishConf', 'publishService', 'utilService', '$filter', 'groups', 'users', 'properties'];
 
 })(angular.module('ov.publish'));